# internal/api/server/server.gen.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package server

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// Defines values for ApplicationService.
const (
	Webserver ApplicationService = "webserver"
)

// Application defines model for Application.
type Application struct {
	// Name Name of the application
	Name string `json:"name"`

	// Service Service of the application
	Service ApplicationService `json:"service"`

	// Tier Policy Tier of the application
	Tier *string `json:"tier,omitempty"`

	// Zones Zones of the application
	Zones *[]string `json:"zones,omitempty"`
}

// ApplicationService Service of the application
type ApplicationService string

// ApplicationList defines model for ApplicationList.
type ApplicationList = []Application

// DeclaredVm defines model for DeclaredVm.
type DeclaredVm struct {
	// Cpu Number of CPU cores
	Cpu int `json:"cpu"`

	// CreatedAt CreatedAt
	CreatedAt *string `json:"createdAt,omitempty"`

	// DnsName DNS Name
	DnsName *string `json:"dnsName,omitempty"`

	// Env Environment (dev, staging, prod)
	Env string `json:"env"`

	// Gateway Gateway
	Gateway *string `json:"gateway,omitempty"`

	// Id Declared VM ID
	Id *string `json:"id,omitempty"`

	// IpAddress IP Address
	IpAddress *string `json:"ipAddress,omitempty"`

	// Name Name of the virtual machine
	Name string `json:"name"`

	// Netmask Netmask
	Netmask *string `json:"netmask,omitempty"`

	// Os Operating system
	Os string `json:"os"`

	// Ram RAM in GB
	Ram int `json:"ram"`

	// Region Target region for placement
	Region *string `json:"region,omitempty"`

	// Role Role of the VM
	Role string `json:"role"`

	// TenantId Tenant ID
	TenantId *string `json:"tenantId,omitempty"`
}

// DeclaredVmList defines model for DeclaredVmList.
type DeclaredVmList = []DeclaredVm

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code *int `json:"code,omitempty"`

	// Error Error message
	Error string `json:"error"`
}

// PlacementResponse defines model for PlacementResponse.
type PlacementResponse struct {
	// Host Target host
	Host *string `json:"host,omitempty"`

	// Id Placement ID
	Id *string `json:"id,omitempty"`

	// Message Status message
	Message string `json:"message"`

	// Status Placement status
	Status *string `json:"status,omitempty"`
}

// RequestedVmList defines model for RequestedVmList.
type RequestedVmList = []VM

// VM defines model for VM.
type VM struct {
	// Cpu Number of CPU cores
	Cpu int `json:"cpu"`

	// Env Environment (dev, staging, prod)
	Env string `json:"env"`

	// Name Name of the virtual machine
	Name string `json:"name"`

	// Os Operating system
	Os string `json:"os"`

	// Ram RAM in GB
	Ram int `json:"ram"`

	// Region Target region for placement
	Region *string `json:"region,omitempty"`

	// Role Role of the VM
	Role string `json:"role"`

	// TenantId Tenant ID
	TenantId *string `json:"tenantId,omitempty"`
}

// CreateApplicationJSONRequestBody defines body for CreateApplication for application/json ContentType.
type CreateApplicationJSONRequestBody = Application

// PlaceVMJSONRequestBody defines body for PlaceVM for application/json ContentType.
type PlaceVMJSONRequestBody = VM

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get all applications
	// (GET /applications)
	GetApplications(w http.ResponseWriter, r *http.Request)
	// Create an application
	// (POST /applications)
	CreateApplication(w http.ResponseWriter, r *http.Request)
	// Get all declared virtual machines
	// (GET /declaredvms)
	GetDeclaredVms(w http.ResponseWriter, r *http.Request)
	// Health check
	// (GET /health)
	Health(w http.ResponseWriter, r *http.Request)
	// Place a virtual machine
	// (POST /place/vm)
	PlaceVM(w http.ResponseWriter, r *http.Request)
	// Get all requested virtual machines
	// (GET /requestedvms)
	GetRequestedVms(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Get all applications
// (GET /applications)
func (_ Unimplemented) GetApplications(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create an application
// (POST /applications)
func (_ Unimplemented) CreateApplication(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get all declared virtual machines
// (GET /declaredvms)
func (_ Unimplemented) GetDeclaredVms(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Health check
// (GET /health)
func (_ Unimplemented) Health(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Place a virtual machine
// (POST /place/vm)
func (_ Unimplemented) PlaceVM(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get all requested virtual machines
// (GET /requestedvms)
func (_ Unimplemented) GetRequestedVms(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetApplications operation middleware
func (siw *ServerInterfaceWrapper) GetApplications(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetApplications(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// CreateApplication operation middleware
func (siw *ServerInterfaceWrapper) CreateApplication(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateApplication(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetDeclaredVms operation middleware
func (siw *ServerInterfaceWrapper) GetDeclaredVms(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetDeclaredVms(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// Health operation middleware
func (siw *ServerInterfaceWrapper) Health(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Health(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// PlaceVM operation middleware
func (siw *ServerInterfaceWrapper) PlaceVM(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlaceVM(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetRequestedVms operation middleware
func (siw *ServerInterfaceWrapper) GetRequestedVms(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRequestedVms(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/applications", wrapper.GetApplications)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/applications", wrapper.CreateApplication)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/declaredvms", wrapper.GetDeclaredVms)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/health", wrapper.Health)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/place/vm", wrapper.PlaceVM)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/requestedvms", wrapper.GetRequestedVms)
	})

	return r
}

type GetApplicationsRequestObject struct {
}

type GetApplicationsResponseObject interface {
	VisitGetApplicationsResponse(w http.ResponseWriter) error
}

type GetApplications200JSONResponse ApplicationList

func (response GetApplications200JSONResponse) VisitGetApplicationsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetApplications400JSONResponse Error

func (response GetApplications400JSONResponse) VisitGetApplicationsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type CreateApplicationRequestObject struct {
	Body *CreateApplicationJSONRequestBody
}

type CreateApplicationResponseObject interface {
	VisitCreateApplicationResponse(w http.ResponseWriter) error
}

type CreateApplication201JSONResponse Application

func (response CreateApplication201JSONResponse) VisitCreateApplicationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateApplication400JSONResponse Error

func (response CreateApplication400JSONResponse) VisitCreateApplicationResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetDeclaredVmsRequestObject struct {
}

type GetDeclaredVmsResponseObject interface {
	VisitGetDeclaredVmsResponse(w http.ResponseWriter) error
}

type GetDeclaredVms200JSONResponse DeclaredVmList

func (response GetDeclaredVms200JSONResponse) VisitGetDeclaredVmsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetDeclaredVms400JSONResponse Error

func (response GetDeclaredVms400JSONResponse) VisitGetDeclaredVmsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type HealthRequestObject struct {
}

type HealthResponseObject interface {
	VisitHealthResponse(w http.ResponseWriter) error
}

type Health200Response struct {
}

func (response Health200Response) VisitHealthResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type PlaceVMRequestObject struct {
	Body *PlaceVMJSONRequestBody
}

type PlaceVMResponseObject interface {
	VisitPlaceVMResponse(w http.ResponseWriter) error
}

type PlaceVM200JSONResponse PlacementResponse

func (response PlaceVM200JSONResponse) VisitPlaceVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PlaceVM400JSONResponse Error

func (response PlaceVM400JSONResponse) VisitPlaceVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type PlaceVM500JSONResponse Error

func (response PlaceVM500JSONResponse) VisitPlaceVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type GetRequestedVmsRequestObject struct {
}

type GetRequestedVmsResponseObject interface {
	VisitGetRequestedVmsResponse(w http.ResponseWriter) error
}

type GetRequestedVms200JSONResponse RequestedVmList

func (response GetRequestedVms200JSONResponse) VisitGetRequestedVmsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetRequestedVms400JSONResponse Error

func (response GetRequestedVms400JSONResponse) VisitGetRequestedVmsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Get all applications
	// (GET /applications)
	GetApplications(ctx context.Context, request GetApplicationsRequestObject) (GetApplicationsResponseObject, error)
	// Create an application
	// (POST /applications)
	CreateApplication(ctx context.Context, request CreateApplicationRequestObject) (CreateApplicationResponseObject, error)
	// Get all declared virtual machines
	// (GET /declaredvms)
	GetDeclaredVms(ctx context.Context, request GetDeclaredVmsRequestObject) (GetDeclaredVmsResponseObject, error)
	// Health check
	// (GET /health)
	Health(ctx context.Context, request HealthRequestObject) (HealthResponseObject, error)
	// Place a virtual machine
	// (POST /place/vm)
	PlaceVM(ctx context.Context, request PlaceVMRequestObject) (PlaceVMResponseObject, error)
	// Get all requested virtual machines
	// (GET /requestedvms)
	GetRequestedVms(ctx context.Context, request GetRequestedVmsRequestObject) (GetRequestedVmsResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetApplications operation middleware
func (sh *strictHandler) GetApplications(w http.ResponseWriter, r *http.Request) {
	var request GetApplicationsRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetApplications(ctx, request.(GetApplicationsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetApplications")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetApplicationsResponseObject); ok {
		if err := validResponse.VisitGetApplicationsResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreateApplication operation middleware
func (sh *strictHandler) CreateApplication(w http.ResponseWriter, r *http.Request) {
	var request CreateApplicationRequestObject

	var body CreateApplicationJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateApplication(ctx, request.(CreateApplicationRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateApplication")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateApplicationResponseObject); ok {
		if err := validResponse.VisitCreateApplicationResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetDeclaredVms operation middleware
func (sh *strictHandler) GetDeclaredVms(w http.ResponseWriter, r *http.Request) {
	var request GetDeclaredVmsRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetDeclaredVms(ctx, request.(GetDeclaredVmsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetDeclaredVms")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetDeclaredVmsResponseObject); ok {
		if err := validResponse.VisitGetDeclaredVmsResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// Health operation middleware
func (sh *strictHandler) Health(w http.ResponseWriter, r *http.Request) {
	var request HealthRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.Health(ctx, request.(HealthRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Health")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(HealthResponseObject); ok {
		if err := validResponse.VisitHealthResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PlaceVM operation middleware
func (sh *strictHandler) PlaceVM(w http.ResponseWriter, r *http.Request) {
	var request PlaceVMRequestObject

	var body PlaceVMJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PlaceVM(ctx, request.(PlaceVMRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PlaceVM")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PlaceVMResponseObject); ok {
		if err := validResponse.VisitPlaceVMResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetRequestedVms operation middleware
func (sh *strictHandler) GetRequestedVms(w http.ResponseWriter, r *http.Request) {
	var request GetRequestedVmsRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetRequestedVms(ctx, request.(GetRequestedVmsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetRequestedVms")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetRequestedVmsResponseObject); ok {
		if err := validResponse.VisitGetRequestedVmsResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}